encapsulate atc.Routes
no consumers are actually expecting to interact with a whole rata.Route, they
all need only the names.
