+++ atc/wrappa/api_auth_wrappa_test.go
@@ -109,8 +109,8 @@ var _ = Describe("APIAuthWrappa", func() {
 		BeforeEach(func() {
 			inputHandlers = map[string]http.Handler{}
 
-			for _, route := range atc.Routes {
-				inputHandlers[route.Name] = &stupidHandler{}
+			for _, routeName := range atc.RouteNames() {
+				inputHandlers[routeName] = &stupidHandler{}
 			}
 
 			expectedHandlers = map[string]http.Handler{
